requirements.txt
pywebview>=4.0
requests
